# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Coziyoo API v1
  version: 1.0.0
  description: Versioned API contract for Coziyoo marketplace backend.
servers:
  - url: http://localhost:3000
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName, userType]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                displayName: { type: string }
                userType: { type: string, enum: [buyer, seller, both] }
      responses:
        '201': { description: Registered }
        '400': { description: Validation error }
        '409': { description: Duplicate email/display name }
  /v1/auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Logged in }
        '401': { description: Invalid credentials }
  /v1/auth/refresh:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: Refreshed }
...
  /v1/auth/display-name/check:
    get:
      tags: [auth]
      security: []
      parameters:
        - in: query
          name: value
          required: true
          schema: { type: string }
      responses:
        '200': { description: Display name availability }
  /v1/admin/auth/login:
    post:
      tags: [admin-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Admin logged in }
  /v1/admin/auth/refresh:
    post:
      tags: [admin-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: Refreshed }
...